fix: better handling of seccomp_ctx
Use std::unique_ptr to make sure the context is properly freed if
creation errors, and make sure the pointer is also zeroed when
appropriate. The value in the config struct is still a raw pointer
because fixing ownership properly is too much effort for a codebase that
is distinctly anti-C++ in style.